backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
tavily-python
google-genai
